+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Schema Delete 
   * Schema Toggle 
   * Schema SetFlag
+  * Schema SetFlags
 ## v0.3.2 - Model Package 
   * Commit: 2026-03-31 16:19
   * Model fields